% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_lgbtq}
\alias{palette_lgbtq}
\title{Get one of LGBTQ palettes}
\usage{
palette_lgbtq(name)
}
\arguments{
\item{name}{\code{character(1)} \cr
Name of the flag the colors are based on.}
}
\value{
An \code{lgbtq_palette} object containing a vector of RGB color codes as strings.
}
\description{
Returns a palette object with colors based on a specified LGBTQ
flag. Note: the number of colors vary between palettes!
}
\examples{
palette_lgbtq("bisexual")

}
