% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_lgbtq.R
\name{seq.lgbtq_palette}
\alias{seq.lgbtq_palette}
\alias{seq.lgbtq_palette_hflag_symmetrical}
\alias{seq.lgbtq_palette_hflag_outer}
\title{Generate colour sequence from a palette}
\usage{
\method{seq}{lgbtq_palette}(x, ...)

\method{seq}{lgbtq_palette_hflag_symmetrical}(x, ...)

\method{seq}{lgbtq_palette_hflag_outer}(x, ...)
}
\arguments{
\item{x}{\code{lgbtq_palette()}\cr
Palette to generate sequence from.}

\item{...}{\cr
These dots are for future extensions and must be empty.}
}
\value{
A character vector of colours in flag order.
}
\description{
Transforms a palette into a sequence of colours corresponding to the order
of those colours on the corresponding flag.
}
