% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text_lgbtq}
\alias{text_lgbtq}
\alias{text_lgbtq.character}
\alias{text_lgbtq.cli_ansi_string}
\title{Stylize text in LGBTQ colors}
\usage{
text_lgbtq(
  text,
  palette,
  ...,
  .style = c("stripe", "cycle_char", "cycle_word")
)

\method{text_lgbtq}{character}(
  text,
  palette,
  ...,
  .style = c("stripe", "cycle_char", "cycle_word")
)

\method{text_lgbtq}{cli_ansi_string}(
  text,
  palette,
  ...,
  .style = c("stripe", "cycle_char", "cycle_word")
)
}
\arguments{
\item{text}{\code{character() | cli_ansi_string()}\cr
A character vector to paint over.}

\item{palette}{\code{character(1)}\cr
One of the available palettes to use.}

\item{...}{\cr
These dots are for future extensions and must be empty.}

\item{.style}{\code{character(1)}\cr
The manner in which the colors are applied.}
}
\value{
A CLI ANSI string vector with colors applied, everything else
remaining intact (vector length, string contents, etc.).
}
\description{
Colors the supplied text according to the order of colors on the flag
corresponding to the selected palette.
}
\examples{
string_lgbtq(
  "This message was brought to you by Big Gay",
  "gay"
)

}
