% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pride.R
\name{show_pride}
\alias{show_pride}
\title{List available palettes}
\usage{
show_pride(..., tags = NULL, sizes = NULL)
}
\arguments{
\item{...}{\cr
These dots are for future extensions and must be empty.}

\item{tags}{\code{character()} \cr
Tags to filter by; currently available:
\code{"rainbow"}, \code{"sexuality"}, \code{"gender"}, \code{"subculture"}, \code{"a-spec"}, \code{"galactian"}.}

\item{sizes}{\code{integer()} \cr
Palette sizes to require.}
}
\value{
An \code{lgbtq_palette_list} with matching palettes.
}
\description{
Shows the available LGBTQ palettes, possibly filtered by tags and/or size.
}
\examples{
# By default, all palettes are shown
show_pride()

# To find rainbow palettes of sizes 6 and 8:
show_pride(tags = "rainbow", sizes = c(6, 8))

}
