% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pride.R
\name{show_pride}
\alias{show_pride}
\title{List available palettes}
\usage{
show_pride(tags = NULL, sizes = NULL)
}
\value{
An \code{lgbtq_palette_list} with matching palettes.
}
\description{
Shows the names of available LGBTQ palettes and their metadata.
}
\examples{
show_pride()

}
